progress-tracker/requirements.txt
gunicorn
Django==2.2.4
django-filter
djangorestframework
mysqlclient
